defer loading org
Instead of requiring `org' when `orglink' is loaded, defer doing so
until `orglink-mode' is called.  Loading `org' takes a long time.

Add a kludge to `turn-on-orglink-mode-if-desired' to prevent
`orglink-mode' from being activated in the *scratch* buffer.  If
we didn't do that then `org' would be loaded at startup anyway.

Also declare some functions and variables to quiet the byte-compiler.