require org at load time
Previously `org' was required inside `orglink-mode'.  The idea was that
this would decrease startup time.  But unfortunately it also meant that
the first time `orglink-mode' got enabled in some buffer it ended up
being called nine times.  And I have no idea why.

Also stop avoiding to turn on `orglink-mode' in the "*scratch*" buffer.
Since we have given up on not loading `org' on startup, that too is now
pointless.