orglink-activate-bracket-links: Adjust to Org change
The regexp used to match bracket links no longer has a third submatch,
but using the second submatch instead should work in most cases.

Eventually this library should be refactored the same way as the
respective code in Org has leading up to Org v9.3.  But that is too
involved for a hotfix.