Copy and adjust updated link-highlighting code from Org 9.5
In [1: 995d9d1] we copied the implementation from Org and
then adjusted it in [2: 4e3e6d9] to respect
`orglink-inside-comment-or-docstring-p', and later again
in [3: 657d655] to respect `orglink-match-anywhere'.
In between we ported over minor changes changes to Org's
implementation.

That implementation has now changed significantly and we
copy it over once more, while making our adjustments in
the same commit.

1: 2016-06-06 995d9d14f94f80d7147aef586899be541c1666c0
   add orglink-activate-{bracket,angle,plain}-links

2: 2016-06-06 4e3e6d920a74fd32a57d5722f81293428e9d8a46
   don't look for links in non-doc-string strings

3: 2021-10-06 657d655caf408283cc1c6fcf437f33b25166244b
   orglink-match-anywhere: New option